bf: ZENKO-1005 auto set offest for new locations
New storage locations will have no initial offset for each
partition. This causes problems if the user were to add a
new or fairly new storage location and were to pause the
location. Since pausing causes consumers to unsubscribe
from the crr topic, any partitions that did not receive any
prior data will not be able to determine the LAG for that
partition.

When the location is resumed, since potentially no LAG is
shown, the crr attempt for the new storage location objects
will stay stuck at PENDING.